% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capital_cities}
\alias{capital_cities}
\title{The point locations of Australian capital cities.}
\format{
A data frame with 8 rows and 3 variables:
\describe{
\item{points}{name of cities}
\item{longitude}{location of point in longitude degrees}
\item{latitude}{location of point in latitude degrees}
}
}
\usage{
capital_cities
}
\description{
A dataset containing the longitude and latitude values of Australian
capital cities.
}
\keyword{datasets}
