% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_centroids.R
\name{create_centroids}
\alias{create_centroids}
\title{Create a data frame of longitude and latitude centroids of each polygon}
\usage{
create_centroids(shp_sf, sf_id = NULL, projstring = projstring,
  verbose = FALSE)
}
\arguments{
\item{shp_sf}{an sf object, a data set with a simple feature list column}

\item{sf_id}{a string to indicate the column to identify individual polygons}

\item{projstring}{a string to indicate the projection and epsg}

\item{verbose}{a boolean to indicate whether to show function progress}
}
\value{
a tibble containing longitude and latitude
}
\description{
Create a data frame of longitude and latitude centroids of each polygon
}
