% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_hexagon.R
\name{fortify_hexagon}
\alias{fortify_hexagon}
\title{Creates the points that define a hexagon polygon for plotting}
\usage{
fortify_hexagon(data, hex_size)
}
\arguments{
\item{data}{data frame containing the longitude, latitude of a hexagon centroid}

\item{hex_size}{the distance between the hexagon centroids}
}
\value{
a data frame of the seven points used to draw a hexagon
}
\description{
Creates the points that define a hexagon polygon for plotting
}
\examples{
# Create centroids set
centroids <- create_centroids(tas_sa2, "SA2_5DIG16")
# Create hexagon location grid
grid <- create_grid(centroids = centroids, hex_size = 0.1, buffer_dist = 0.1)
# Allocate polygon centroids to hexagon grid points
hex_allocated <- allocate(centroids = centroids,
hex_grid = grid,
hex_size = 0.1, # same size used in create_grid
filter_dist = 0.25,
focal_points = sugaRbag::capital_cities,
width = 30,
id = "SA2_5DIG16", verbose = TRUE) # same column used in create_centroids

fortify_hexagon(hex_allocated[1,], 0.1)
}
