% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hexmap_aus_3}
\alias{hexmap_aus_3}
\title{A hexagon map data set of Australia, with hexagons of size 0.3 degrees.}
\format{A data frame with 2192 rows and 11 variables:
\describe{
\item{SA2_NAME11}{name of the Statistical Area (Level 2)}
\item{longitude}{location of SA2 in longitude degrees}
\item{latitude}{location of SA2 in latitude degrees}
\item{focal_point}{name of the closest Australian capital city}
\item{longitude1}{location of the closest Australian capital city in
longitude degrees}
\item{latitude1}{location of the closest Australian capital city in
latitude degrees}
\item{focal_dist}{distance in metres between the polygon centroid and closest capital city}
\item{focal_angle}{angle between the closest capital city and polygon centroid}
\item{hex_long}{location of the hexagon grid point in    longitude degrees}
\item{hex_lat}{location of the hexagon grid point in latitude degrees}
\item{hex_id}{id of the hexagon grid point}
}}
\usage{
hexmap_aus_3
}
\description{
A dataset containing information regarding the closest capital city, and the
allocated hexagon grid point for each Statistical Area (Level 2) in
Australia.
}
\keyword{datasets}
