% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homeless}
\alias{homeless}
\title{The amount of homeless people in each Local Government Areas in 2016.}
\format{A data frame with 545 rows and 2 variables:
\describe{
\item{homeless}{amount of homeless people}
\item{LGA_NAME16}{name of the Local Government Area}
}}
\usage{
homeless
}
\description{
A data frame of the Local Government Area names and amount of homeless.
}
\keyword{datasets}
