% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sugarbag.R
\name{make_sugarbag_df}
\alias{make_sugarbag_df}
\title{Create a dataframe of hexagon locations for use in
geom_sugarbag(). This function is an opinionated wrapper around
various other functions in the sugarbag package.}
\usage{
make_sugarbag_df(shp, hex_size = 0.2, hex_width = 50)
}
\arguments{
\item{shp}{A shapefile of class \code{"sf"}}

\item{hex_size}{Default 0.2. See \code{?allocate}}

\item{hex_width}{Default 50. See \code{?allocate} (argument \code{width})}
}
\description{
Create a dataframe of hexagon locations for use in
geom_sugarbag(). This function is an opinionated wrapper around
various other functions in the sugarbag package.
}
