% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_to_tibble.R
\name{sfc_to_tibble}
\alias{sfc_to_tibble}
\title{Convert a simple features tibble to tibble for plotting. This will contain
individual points for plotting the polygon, indicating the longitude and
latitude, order of points, if a hole is present, the piece, id and group.}
\usage{
sfc_to_tibble(sfc_df)
}
\arguments{
\item{sfc}{a simple features table with a geometry column}
}
\value{
a tibble point of long lat points used to plot polygons
}
\description{
Convert a simple features tibble to tibble for plotting. This will contain
individual points for plotting the polygon, indicating the longitude and
latitude, order of points, if a hole is present, the piece, id and group.
}
