% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugarbag-package.R
\docType{package}
\name{sugarbag-package}
\alias{sugarbag}
\alias{sugarbag-package}
\title{sugarbag: Create Tessellated Hexagon Maps}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create a hexagon tile map display from spatial polygons. Each polygon is represented by a hexagon tile, placed as close to it's original centroid as possible, with a focus on maintaining spatial relationship to a focal point. Developed to aid visualisation and analysis of spatial distributions across Australia, which can be challenging due to the concentration of the population on the coast and wide open interior.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://srkobakian.github.io/sugarbag/}
  \item \url{https://github.com/srkobakian/sugarbag}
}

}
\author{
\strong{Maintainer}: Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID}) [thesis advisor]

Authors:
\itemize{
  \item Stephanie Kobakian (\href{https://orcid.org/0000-0002-4616-0048}{ORCID})
}

Other contributors:
\itemize{
  \item Matt Cowgill [contributor]
}

}
\keyword{internal}
