% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tas_lga}
\alias{tas_lga}
\title{The polygons of Tasmanian Local Government Areas in 2016.}
\format{
A simple features data frame with 39 rows and 6 variables:
\describe{
\item{lga_code_2016}{code for the Local Government Area}
\item{lga_name_2016}{name of the Local Government Area}
\item{ste_code_2016}{code for the state containing the Local Government Area}
\item{ste_name_2016}{name of the state containing the Local Government Area}
\item{areasqkm_2016}{area contained in the polygon}
\item{geometry}{describes where on Earth the polygon is located}
}
}
\usage{
tas_lga
}
\description{
A simple features dataset containing the polygons for all Australian LGAs in
2016.
}
\keyword{datasets}
