% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tas_lga_hexctr}
\alias{tas_lga_hexctr}
\title{The hexagon centres for polygons of Tasmanian Local Government Areas in 2016.}
\format{
A simple features data frame with 39 rows and 6 variables:
\describe{
\item{lga_code_2016}{code for the Local Government Area}
\item{longitude, latitude}{polygon centroid}
\item{points, focal_longitude, focal_latitude, focal_dist, focal_angle}{Focal point (capital city) information used for each polygon/hexagon}
\item{rownumber}{row number, in case it can be useful}
\item{hex_long, hex_lat, hex_id}{hexagon centre and id}
}
}
\usage{
tas_lga_hexctr
}
\description{
A tibble dataset containing the processed data for all Australian LGAs in
2016. Each point corresponds to hexagon centre.
}
\keyword{datasets}
