% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tas_sa2}
\alias{tas_sa2}
\title{The polygons of Tasmanian Statistical Areas in 2016.}
\format{
A simple features data frame with 99 rows and 15 variables:
\describe{
\item{sa2_main_2016}{complete code of the Statistical Area}
\item{sa2_5dig_2016}{simple code for the Statistical Area}
\item{sa2_name_2016}{name of the Statistical Area}
\item{sa3_code_2016}{code for the SA3 containing the Statistical Area}
\item{sa3_name_2016}{name of the SA3 containing the Statistical Area}
\item{sa4_code_2016}{code for the SA4 containing the Statistical Area}
\item{sa4_name_2016}{name of the SA4 containing the Statistical Area}
\item{gcc_code_2016}{code for the Greater Capital City region containing
the Statistical Area}
\item{gcc_name_2016}{name of the Greater Capital City region containing
the Statistical Area}
\item{ste_code_2016}{code for the state containing the Statistical Area}
\item{ste_name_2016}{name of the state containing the Statistical Area}
\item{areasqkm_2016}{area contained in the polygon}
\item{id}{distinguishes SA2 regions}
\item{population}{amount of people living within the region}
\item{sa2_code_2016}{code of the Statistical Area}
}
}
\usage{
tas_sa2
}
\description{
A simple features dataset containing the polygons for all Tasmanian SA2s in
2016.
}
\keyword{datasets}
