% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Allocate polygon centroids to hexagons in a grid}
\usage{
allocate(centroids, grid, radius)
}
\arguments{
\item{centroids}{a data frame with centroids of non empty polygons}

\item{grid}{a data frame containing points to allocate hexagons}

\item{radius}{distance between each grid point}
}
\description{
Chooses a hexagon centroid for each polygon in the shape file, from a grid
spanning the longitudes and latitudes in the expanded bounding box.
}
