% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Allocate polygon centroids to hexagons in a grid}
\usage{
allocate(centroids, hex_grid, hex_size, filter_dist, focal_points,
  show_progress, id)
}
\arguments{
\item{centroids}{a data frame with centroids of non empty polygons}

\item{hex_grid}{a data frame containing all possible hexagon points}

\item{hex_size}{distance between each grid point}

\item{filter_dist}{distance around centroid to consider hex points}
}
\value{
a data frame of one allocation
}
\description{
Chooses a hexagon centroid for each polygon in the shape file, from a grid
spanning the longitudes and latitudes in the expanded bounding box.
}
