% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Allocate polygon centroids to hexagons in a grid}
\usage{
allocate(centroids, hex_grid, hex_size, filter_dist, focal_points = NULL,
  verbose, id)
}
\arguments{
\item{centroids}{a data frame with centroids of non empty polygons}

\item{hex_grid}{a data frame containing all possible hexagon points}

\item{hex_size}{a float value in degrees for the diameter of the hexagons}

\item{filter_dist}{distance around centroid to consider grid points}

\item{focal_points}{a data frame of reference locations when allocating
hexagons, capital cities of Australia are used in the example}

\item{verbose}{a boolean to indicate whether to show polygon id}

\item{id}{a string to indicate the column to identify individual polygons}
}
\value{
a data frame of one allocation
}
\description{
Chooses a hexagon centroid for each polygon in the shape file, from a grid
spanning the longitudes and latitudes in the expanded bounding box.
}
