% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_focal_point.R
\name{closest_focal_point}
\alias{closest_focal_point}
\title{For the polygon provided, find the closest focal point in the set provided}
\usage{
closest_focal_point(centroid, focal_points = as.matrix(focal_points))
}
\arguments{
\item{focal_points}{}
}
\value{
data frame containing the name and location of the closest focal
}
\description{
For one row of an sf data frame, calculate the distance to the closest focal
point. Return the name of the focal point, and the angle between focal point and centroid.
}
