% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid of evenly spaced points to allow hexagons to tesselate}
\usage{
create_grid(centroids, bbox, hex_size, buffer_dist, verbose = FALSE)
}
\arguments{
\item{centroids}{data frame of centroids to be allocated}

\item{bbox}{table with columns minimum and maximum and rows
containing longitude and latitude values}

\item{hex_size}{a float value in degrees for the diameter of the hexagons}

\item{buffer_dist}{distance to extend beyond the geometry provided}

\item{verbose}{a boolean to indicate whether to show function progress}
}
\value{
grid
}
\description{
This function takes the bounding box of a group of polygons, or a specific
table of minimum and maximum longitudes and latitudes to create points
for each polygon to be allocated to that will tesselate into hexagons.
}
