% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexmap.R
\name{create_hexmap}
\alias{create_hexmap}
\title{create_hexmap
Allocates each polygon in a shape file to a grid point to create a map of
tesselated hexagons. The spatial relationships of areas are preserved while
the geographic chape of each area is lost.}
\usage{
create_hexmap(shp, bbox = NULL, radius = "auto", focal_points = NULL,
  expand_grid = 0.1)
}
\arguments{
\item{shp}{a file containing polygons to allocate to a hexmap}

\item{bbox}{table with columns minimum and maximum and rows
containing longitude and latitude values}

\item{radius}{a float value in degrees for the diameter of the hexagons}

\item{expand_grida}{a float value to expand the grid beyond the minimum
and maximum longitudes and latitudes}
}
\description{
create_hexmap
Allocates each polygon in a shape file to a grid point to create a map of
tesselated hexagons. The spatial relationships of areas are preserved while
the geographic chape of each area is lost.
}
\examples{
#shp <- "C:/Users/steff/Documents/atlas/hexmap/data/LGA_2011_AUST/LGA_2011_AUST.shp"


}
