% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexmap.R
\name{create_hexmap}
\alias{create_hexmap}
\title{Create a tesselated hexagon map from a set of polygons}
\usage{
create_hexmap(shp = NULL, shp_path = NULL, sf_id = NULL,
  buffer_dist = NULL, hex_size = NULL, filter_dist = NULL,
  width = 15, focal_points = NULL, export_shp = FALSE,
  verbose = FALSE)
}
\arguments{
\item{shp}{a shape file}

\item{shp_path}{character string location of shape file}

\item{sf_id}{name of a unique column that distinguishes areas}

\item{buffer_dist}{distance in degrees to extend beyond the geometry provided}

\item{hex_size}{a float value in degrees for the diameter of the hexagons}

\item{filter_dist}{amount of hexagons around centroid to consider for allocation}

\item{width}{the angle used to filter the grid points around a centroid}

\item{focal_points}{a data frame of reference locations when allocating
hexagons, capital cities of Australia are used in the example}

\item{export_shp}{export the simple features set}

\item{verbose}{a boolean to indicate whether to show function progress}
}
\value{
a data set containing longitude and latitude of allocated hexagon
points for each non null geometry passed in the shape file
}
\description{
Allocates each polygon in a shape file to a grid point to create a map of
tesselated hexagons. The spatial relationships of areas are preserved while
the geographic chape of each area is lost.
}
\examples{
\dontrun{

data(tas_sa2)
hexmap <- create_hexmap(shp = tas_sa2, sf_id = "SA2_5DIG16",
filter_dist = 10, export_shp = FALSE,
focal_points = sugaRbag::capital_cities, verbose = TRUE)
}

}
