% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexmap.R
\name{create_hexmap}
\alias{create_hexmap}
\title{Create a tesselated hexagon map from a set of polygons}
\usage{
create_hexmap(shp_path, buffer_dist = NULL, hex_size = "auto",
  filter_dist = 1000, focal_points = NULL, export_shp = FALSE)
}
\arguments{
\item{shp_path}{character vector location of shape file, extension .shp}

\item{hex_size}{a float value in degrees for the diameter of the hexagons}

\item{focal_points}{a data frame of reference locations when allocating
hexagons, capital cities of Australia are used in the example}

\item{export_shp}{export the simple features set}

\item{buffer}{a float distance in degrees to exapnd hexagon grid}
}
\value{
a data set containing longitude and latitude of allocated hexagon
points for each non null geometry passed in the shape file
}
\description{
Allocates each polygon in a shape file to a grid point to create a map of
tesselated hexagons. The spatial relationships of areas are preserved while
the geographic chape of each area is lost.
}
\examples{
\dontrun{
# change to sa2_2011 on github\\
shp_path <- system.file("data","sa2_2011.Rda", package = "sugaRbag")
load(system.file("data","capital_cities.Rda", package = "sugaRbag"))

hexmap <- create_hexmap(shp_path = shp_path, buffer_dist = NULL, hex_size = "auto", export_shp = FALSE,
focal_points = capital_cities)
}

}
