% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexmap.R
\name{create_hexmap}
\alias{create_hexmap}
\title{Create a tesselated hexagon map from a set of polygons}
\usage{
create_hexmap(shp = NULL, shp_path = NULL, sf_id = NULL,
  buffer_dist = NULL, hex_size = NULL, filter_dist = NULL,
  width = 30, focal_points = NULL, export_shp = FALSE,
  verbose = FALSE)
}
\arguments{
\item{shp}{a shape file}

\item{shp_path}{character string location of shape file}

\item{buffer_dist}{distance in degrees to extend beyond the geometry provided}

\item{hex_size}{a float value in degrees for the diameter of the hexagons}

\item{filter_dist}{amount of hexagons around centroid to consider for allocation}

\item{focal_points}{a data frame of reference locations when allocating
hexagons, capital cities of Australia are used in the example}

\item{export_shp}{export the simple features set}

\item{verbose}{a boolean to indicate whether to show function progress}
}
\value{
a data set containing longitude and latitude of allocated hexagon
points for each non null geometry passed in the shape file
}
\description{
Allocates each polygon in a shape file to a grid point to create a map of
tesselated hexagons. The spatial relationships of areas are preserved while
the geographic chape of each area is lost.
}
\examples{
\dontrun{

shp_path <- system.file("data","sa2_2011.Rda", package = "sugaRbag")
load(system.file("data","capital_cities.Rda", package = "sugaRbag"))

create_hexmap(shp_path = shp_path, sf_id = "SA2_NAME11", buffer_dist = NULL,
filter_dist = 10, hex_size = "auto", export_shp = FALSE,
focal_points = capital_cities, verbose = TRUE)
}

}
