% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_grid_points.R
\name{filter_grid_points}
\alias{filter_grid_points}
\title{Filter full set of grid points for those within range of original point}
\usage{
filter_grid_points(fgrid, folong, folat, focal_distance = NULL,
  filter_distance)
}
\arguments{
\item{fgrid}{complete grid of hexagon centroids}

\item{folong}{longitude of centroid of polygon}

\item{folat}{latitude of centroid of polygon}

\item{filter_distance}{a distance in metres, used as a boundary to filter
the hexagon centroids considered for each polygon centroid to be allocated.}

\item{focal_points}{a tibble of focal locations, an optional argument that
allows allocation of polygons to hexagon centroids in ascending order of
the distance to the closest focal point. It also filters the grid points to
those within a 30 degree range of the angle from focal point to centroid.
The default "capitals" uses the locations of the Australian capital cities
as focal points.}
}
\value{
a tibble of filtered grid points
}
\description{
Takes only the closest available gridpoints as possible hexagon centroids
to allocate polygons.
}
\examples{


}
