% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_hexagon.R
\name{fortify_hexagon}
\alias{fortify_hexagon}
\title{Creates the points that define a hexagon polygon for plotting}
\usage{
fortify_hexagon(data, sf_id, hex_size)
}
\arguments{
\item{data}{a data frame created by the allocate function}

\item{sf_id}{a string to indicate the column to identify individual polygons}

\item{hex_size}{a float value in degrees for the diameter of the hexagons}
}
\value{
a data frame of the seven points used to draw a hexagon
}
\description{
Creates the points that define a hexagon polygon for plotting
}
\examples{
# same column is used in create_centroids
fortify_hexagon(data = tas_lga_hexctr, sf_id = "lga_code_2016", hex_size = 0.2)
}
