% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_sfc.R
\name{fortify_sfc}
\alias{fortify_sfc}
\title{Convert a simple features tibble to tibble for plotting.}
\usage{
fortify_sfc(sfc_df, keep = NULL)
}
\arguments{
\item{sfc_df}{a simples features data set}

\item{keep}{ratio of points to keep}
}
\value{
a tibble point of long lat points used to plot polygons
}
\description{
This will contain individual points for plotting the polygon, indicating the
longitude and latitude, order of points, if a hole is present, the piece, id
and group.
}
