% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sugarbag.R
\name{geom_sugarbag}
\alias{geom_sugarbag}
\title{geom_sugarbag}
\usage{
geom_sugarbag(
  mapping = NULL,
  data = NULL,
  stat = "sugarbag",
  position = "identity",
  ...
)
}
\description{
geom_sugarbag
}
\examples{
\donttest{
library(ggplot2)
}
}
