% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sugarbag.R
\name{geom_sugarbag}
\alias{geom_sugarbag}
\title{geom_sugarbag}
\usage{
geom_sugarbag(
  mapping = NULL,
  data = NULL,
  stat = "sugarbag",
  position = "identity",
  ...
)
}
\description{
geom_sugarbag
}
\examples{

tas_lga \%>\% 
  ggplot(aes(fill = lga_name_2016)) +
  geom_sf(alpha = 0.1) +
  geom_sugarbag(aes(geometry = geometry)) +
  theme(legend.position = "none")

}
