% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shape.R
\name{read_shape}
\alias{read_shape}
\title{Read in the shape file as sf object}
\usage{
read_shape(shp_path = NULL, simplify = NULL, epsg = NULL,
  projstring = NULL)
}
\arguments{
\item{shp_path}{character vector location of shape file, extension .shp}

\item{simplify}{boolean to determine whether to simplify the shape file
using rmapshaper}

\item{epsg}{the four character string to indicate the CRS}

\item{projstring}{a string to indicate the projection and epsg}
}
\value{
an sf data frame, with a column of non null geometries
}
\description{
read_shape
}
