% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugarbag-package.R
\docType{package}
\name{sugarbag-package}
\alias{sugarbag}
\alias{sugarbag-package}
\title{sugarbag: Create Tessellated Hexagon Maps}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Create a hexagon tilegram from spatial polygons. Each polygon is
    represented by a hexagon tile, placed as close to it's original centroid
    as possible, with a focus on maintaining spatial relationship to a focal
    point. Developed to aid visualisation and analysis of spatial distributions
    across Australia, which can be challenging due to the concentration of the
    population on the coast and wide open interior.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://srkobakian.github.io/sugaRbag/}
  \item \url{https://github.com/srkobakian/sugaRbag}
}

}
\author{
\strong{Maintainer}: Stephanie Kobakian \email{stephanie.kobakian@gmail.com}

Authors:
\itemize{
  \item Dianne Cook [thesis advisor]
}

}
\keyword{internal}
